Flexible timecol in ScanImages
This commit represents changes made during the January 2018 beamtime at SSRL 11-2, where SPEC was saving .csv files with messed up column names. ScanImages was adjusted to get both a timestamp and time variable from the absolute times saved by that beamline at that time. It proved later, though, to be unstable, and was reverted to the previous commit.