Stabilize, implement timezones, improve metadata storage
The problems from the previous merge seem to have been relatively small. They are fixed, so that XRD plots as well as EC_Xray plots work again, but keeping the better-organizned ScanImages.__init__() and abstimecol functionality. As well of course as the use of epoch time and the improved synchronize function. Both of the latter are taken a step further here. synchronize now saves the metadata from the original files, úsing the original key prefixed by an underscore, as a dictionary indexed by the combining number nd. And, check it out! ...

Timezones now provide absolute truth to epoch times! When a tstamp is calculated from a timestamp, it uses an easily input timezone (defaulting to the local timezone). This enables synchronization accross timezones or dst, which might be a rare necessity, but happens to be essential for analyzing our November beamline data. So... big developments towards making all this code something I'll be proud to publish and distribute.