Read metadata from .raw.pdi
Pilatus objects now get selected metadata from the pdi files. This is especially usefull, because I got Tim Dunn to add a timestamp in epoch time at the end of the .pdi files. ScanImages objects use this time for their own tstamp and relative time by default. This makes synchronization easy (to the accuracy of the spec and EC computers' clocks), even without using the triggers.
Also, added the demo .html's so that I can use github to link to them.